EDIT MISSION COMPONENT
